move Nav into common components
